src/java/main/br/com/rpires/domain/Cliente.java
/**
 * 
 */
package br.com.rpires.domain;

/**
 * @author rodrigo.pires
 *
 */
public class Cliente {

	private Long id;
	private Long cpf;
	private Integer idade;
	private String nome;
	private Long idProduto;

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}
	public String getNome() {
		return nome;
	}

	public void setNome(String nome) {
		this.nome = nome;
	}
	public Long getCpf() {
		return cpf;
	}

	public void setCpf(Long cpf) {
		this.cpf = cpf;
	}

	public Integer getIdade() {
		return idade;
	}

	public void setIdade(Integer idade) {
		this.idade = idade;
	}

	public Long getIdProduto() {
		return idProduto;
	}

	public void setIdProduto(Long idProduto) {
		this.idProduto = idProduto;
	}
}
